Refactor archives
